fix #95. Zoom available in render.

+++ source/core/position.hpp
@@ -34,7 +34,7 @@ public:
     Point operator+(const Point& other) const { return Point( _x + other._x, _y + other._y ); }
     Point operator-(const Point& other) const { return Point( _x - other._x, _y - other._y ); }
     Point operator -() const { return Point( -_x, -_y); }
-    Point operator*(int mul) const{ return Point( _x*mul, _y*mul ); }
+    Point operator*(float mul) const{ return Point( _x*mul, _y*mul ); }
 
     void setX( const int nx ) { _x = nx; }
     void setY( const int ny ) { _y = ny; }
